requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
flask==2.3.3
pytz==2023.3 